JsonTaskBookStorage::moveTaskBook(): only move when source file exists
When the current storage file does not exist (which can happen when the
application is run for the first time), `moveTaskBook()` will throw an
error as it attempts to move an non-existent file.

So, let's teach `moveTaskBook()` to not move the file if the source file
does not exist in the first place.

+++ src/main/java/seedu/address/storage/JsonTaskBookStorage.java
@@ -89,8 +89,10 @@ public class JsonTaskBookStorage implements TaskBookStorage {
         if (FileUtil.isFileExists(newFile)) {
             throw new IOException(newFilePath + " already exists.");
         }
-        Files.createParentDirs(newFile);
-        Files.move(file, newFile);
+        if (FileUtil.isFileExists(file)) {
+            Files.createParentDirs(newFile);
+            Files.move(file, newFile);
+        }
         filePath = newFilePath;
     }
